requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.1
decorator==4.4.2
future==0.18.2
idna==2.9
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.1
python-dotenv==0.13.0
python-telegram-bot==12.7
requests==2.23.0
six==1.15.0
toml==0.10.2
tornado==6.0.4
urllib3==1.25.9
